Solve cyclic dependencies
The package list now has a list of forced packages which are added to
the beginning of the list. A bit of syntax allows to force package
installation etc.

This way packages like freetype2 and harfbuzz are installed in a
specific order, such that we shouldn't end up in an endless loop in
yaourt anymore.